Added a command to view more from the uploader.
Typing "1 author" at the vid prompt will have youtube-viewer display
uploads by the author of the first video on the list. It is also
possible to view playlists instead of uploads with "1 author playlists".

Both of these commands have a one letter alias (a for author, p for
playlists). I intend to add brief aliases for each of the available
commands in the near future.