Fixed minor compatibility issue with yv 3.0.8.
Unfortunately, this comes at the expense of compatibility with previous
versions of yv. Previously, we could just pass the '--non-interactive'
option to yv and it would not spawn a prompt. That option is now
replaced by a line in the configuration file ('interactive => 0') with a
corresponding '--interactive' flag.

ytsm now warns the user if 'interactive => 0' is not set, and removes
the --non-interactive argument from certain parts of the script. It
should still work with older versions, albeit with the same problems as
'interactive => 1' on newer versions.