Partial prettification of cronjob, disable quvirc
So it turns out that man quvi, one of the shortest man pages I have ever
seen, spends a couple lines on QUVI_CONFIG. Not sure how I missed that.
Both of the issues mentioned in the previous commit message have been
fixed. QUVI_CONFIG now points to /dev/null, which produces a warning on
stderr. That isn't a problem because we redirect quvi's stderr to
/dev/null anyway for the sake of formatting.