Experimental change from quvi to youtube-dl
ytsm-cronjob might be better off moving to yt-dl than to quvi 0.9. The
only issue I see is that yt-dl doesn't seem to have a mechanism to
bypass the user's youtube-dl.conf. It might be possible to do something
like 'HOME="$some_other_dir" youtube-dl http...', but I will have to
investigate this further.