Added tools to manipulate the subs file.
These tools are composed of an edit button, an addsub command, and a
delsub command, all within the ytsm frontend. There is error checking
for the one time we use sed, but we haven't gone full visudo on the
user by checking the grammar of the file and whatnot.

If $EDITOR is not set in the edit command, we assume EDITOR=vi. POSIX
doesn't seem to have any official literature on the topic, but given
that the options are either vi or ed, this seems like the best idea.