Added comment parsing, better error checking for as and ds.
Comments in the subs file start with # with optional preceding
whitespace. Comments can either be placed at the end of a line or on a
line of their own. ytsm doesn't really care. The sort option could mess
up comments with their own lines, so the default answer to that question
is now no, and the question warns the user of this limitation.

The new error checking is to make sure that user doesn't answer "s" to a
y/n prompt. It alerts the user that changes have not been made if an
answer other than y, n, or blank input (default answer) is detected.