ytsm-cronjob now reads the real upload date
No more faking it with the current system date. This introduces a few
minor caveats, the biggest of which is that if a new video is for some
reason datestamped a month ago, it won't show up anywhere near the top
of the feed, much less among the new items. The other thing is that the
'today' feature is now essentially useless, as it is mostly replaced by
the 'new' feature in functionality. This will become a 'yesterday'
button that does what the old 'today' button did whenever I next have
time to work on it.

I've introduced some error checking for the new title/date mechanism. If
it doesn't detect the date from youtube-dl, it falls back on system
time. If yt-dl/ytsm can't get the title, it will substitute it for a
brief message telling us that there was a problem.

Also, I deliberately didn't disable the youtube-dl.conf because the
--get-thing options are supposedly simulations, which should skip most
undesirable actions like auto downloading or audio ripping.