today is now yesterday.
The today command is mostly functionally replaced by the yesterday
command. It now prints videos that are uploaded yesterday, or today if
one video uses the new fallback date method (which is the same as the
old method).

I also cleaned up a comment I missed in the POSIX update and added a
comment regarding the escape character that I thought I made but
apparently only went into the commit log.