Fixed a bug in the output stage.
The math done to find $TOTAL_PAGES assumed that the output would round
up to the nearest page. dash/bash have purely integer-based built-in
math, so numbers will always round down. The fix is a simple algebraic
solution: add to the numerator a value equal to one less than the
denominator. This solutions has been documented in a comment block.