Added 'youtube-viewer --resolution audio' support for yv 3.0.9+.
In the event that having an extra window floating around eating up CPU
and GPU cycles is inconvenient, yv does audio only since 3.0.9. ytsm now
accounts for that.

Also changed the help screens from big chunks of echo commands to a
single 'cat << EOF' that reads the text directly out of the script. It's
probably a tiny bit heavier than the shell's built-in echo, but it makes
maintenance on the help screens much, much easier.

I really need to find a reliable way to enable and disable features
based on the youtube-viewer version. That would solve the --interactive
with < 3.0.7 problem as well as the --resolution audio with < 3.0.9
problem. We'll end up with a huge pile of inconsistently implemented
features by the time yv 3.2 is out if we don't, and I really don't want
to force people to follow bleeding edge youtube-viewer if something
major changes.