Improve compatibility with other shells.
A change to the way prompts are presented gives us compatibility with
ksh, mksh, and zsh in addition to bash and dash. All of these shells are
faster than bash is.

As I said in the readme, dash is tied with ksh for the fastest, but
because ytsm-cronjob has a few issues with ksh, dash is still my
recommendation.