Fix a big, obvious bug in as's sort prompt.
In the previous commit, I rewrote as's read line incorrectly, which
effectively made it so that the subs file would never be sorted,
regardless of how the user answers the prompt.

Also tweaked the formatting on ds's prompt. It didn't add a space after
the [yN], which made it look ugly.