Fix a date related bug in ytsm-cronjob.
Two separate variables were both named $DATE, which caused transient
breakage in the filename datestamp, which trickled down to the symlink
process, and the ytsm client as well. This bug seems like it was
probably introduced with commit 5ca21da7 back in December, but for some
reason everything worked fine for me until just a few days ago.